moved `WaitOne` out of `UseConnection`, fixes #9
If there are no jobs to fetch, `SQLiteJobQueue` blocks the thread for
`QueuePollInterval` timespan. And it does that inside `UseConnection`,
therefore `ReleaseDbWriteLock` is not called for that timespan (15
seconds by default). Which means that we cannot write into the database.

The fix is to release db write lock as soon as possible by moving the
`WaitOne` out of the `UseConnection`.